Linux: Mount /dev/shm in child mount namespace
Mounting /dev/shm as part of command execution in the child mount
namespace provides better isolation than mounting it as part of device
installation.
